+++ apps/dashboard/src/app/team/[team_slug]/(team)/~/usage/page.tsx
@@ -1,7 +1,12 @@
+import { ChakraProviderSetup } from "@/components/ChakraProviderSetup";
+import { SettingsUsagePage } from "./UsagePage";
+
 export default function Page() {
   return (
-    <div className="h-full py-6 container flex items-center justify-center">
-      <h1 className="text-4xl tracking-tighter text-muted-foreground">Usage</h1>
-    </div>
+    <ChakraProviderSetup>
+      <div className="container pt-8 pb-10">
+        <SettingsUsagePage />
+      </div>
+    </ChakraProviderSetup>
   );
 }
